InputContainer.test.jsx 작성
